add delete data from certain day

+++ src/App.js
@@ -1,29 +1,56 @@
 import './App.css';
-import {useEffect, useState} from 'react';
+import {useState} from 'react';
 import LineChart from './components/LineChart';
-import DarkModeToggle from './components/DarkModeToggle';
-import {
-  BrowserRouter as Router,
-  Routes,
-  Route,
-  Navigate,
-  BrowserRouter,
-} from "react-router-dom";
-import {Outlet, Link} from "react-router-dom";
+import stringifyDate from './components/FormatDate'
 
+
+const replaceData = (date,weight,setWeightData) => {
+    const existingDataString = localStorage.getItem('weightData')
+    const existingData = existingDataString ? JSON.parse(existingDataString) : [];
+
+    const indexToUpdate = existingData.findIndex(item => item.date === date);
+    if(indexToUpdate !== -1){
+      existingData.splice(indexToUpdate,1)
+    }
+    if(weight === ''){
+      localStorage.setItem('weightData', JSON.stringify(existingData));
+
+      setWeightData({
+        labels: existingData.map((data) => data.date),
+        datasets: [{
+          label:"Paino",
+          data: existingData.map((data)=> data.weight),
+        }]
+      })
+      return;
+    }
+    const newDataPair = {date: date, weight: weight}
+    const newData = [...existingData, newDataPair];
+
+    newData.sort((a, b) => { 
+      const dateA = new Date(a.date.split('-').reverse().join('-'));
+      const dateB = new Date(b.date.split('-').reverse().join('-'));
+    
+      return dateA - dateB;
+    });
+   
+    localStorage.setItem('weightData', JSON.stringify(newData));
+
+    setWeightData({
+      labels: newData.map((data) => data.date),
+      datasets: [{
+        label:"Paino",
+        data: newData.map((data)=> data.weight),
+      }]
+    })
+}
 
 const DatePicker = (props) => {
-
-
   const handleChange = (e) => {
-   
-    const year = new Date(e.target.value).getFullYear();
-    const month = (new Date(e.target.value).getMonth() + 1).toString().padStart(2,'0')
-    const day = new Date(e.target.value).getDate().toString().padStart(2,'0')
-    const formattedDate = `${day}-${month}-${year}`
-
-    props.setDate(formattedDate);
-
+    if(e.target.value !== ''){
+      const stringDate = stringifyDate(new Date(e.target.value));
+      props.setDate(stringDate);
+    }
   };
   return(
       <div>
@@ -39,103 +66,72 @@ const DatePicker = (props) => {
 const WeightForm = (props) => {
   const handleSubmit = (event) => {
     event.preventDefault();
-
-    const newDataPair = {date: props.date, weight: props.weight}
-    const existingDataString = localStorage.getItem('weightData')
-    const existingData = existingDataString ? JSON.parse(existingDataString) : [];
-    const indexToUpdate = existingData.findIndex(item => item.date === props.date);
-    if(indexToUpdate !== -1){
-      existingData.splice(indexToUpdate,1)
-    }
-    const newData = [...existingData, newDataPair];
-    
-
-
-    newData.sort((a, b) => { 
-      const dateA = new Date(a.date.split('-').reverse().join('-'));
-      const dateB = new Date(b.date.split('-').reverse().join('-'));
-    
-      return dateA - dateB;
-    });
-   
-    localStorage.setItem('weightData', JSON.stringify(newData));
-
-    const dataString = localStorage.getItem('weightData')
-    const data = dataString ? JSON.parse(dataString) : [];
-
-
-
-    props.setWeightData({
-      labels: data.map((data) => data.date),
-      datasets: [{
-        label:"Paino",
-        data: data.map((data)=> data.weight),
-      }]
-    })
-    
-
+    replaceData(props.date, props.weight, props.setWeightData);
   }
   return(
-    <>
-    <form onSubmit={handleSubmit}>
-    Paino: <input value = {props.weight} onChange = {(e) => props.setWeight(e.target.value)}/> kg
-    <br/><br/>
-    <button type="submit">Lähetä</button>
-    </form>
-    </>
+    <div>
+      <form onSubmit={handleSubmit}>
+      Paino: <input value = {props.weight} onChange = {(e) => props.setWeight(e.target.value)}/> kg
+      <br/><br/>
+      <button type="submit">Lähetä</button>
+      </form>
+    </div>
+    
   )
 }
-
-const parseDate = (dateString) => {
-  const [day, month, year] = dateString.split('-').map(Number);
-  return new Date(year, month - 1, day);
-};
-
-
 
 const ClearButton = (props) => {
   const handleClear = (event) =>{
     event.preventDefault();
-    localStorage.setItem('weightData', []);
-    props.setWeightData([]);
+    if(window.confirm("Haluatko varmasti poistaa kaikki painotiedot?")){
+      localStorage.setItem('weightData', []);
+      props.setWeightData([]);
+    }
   }
   return(
     <div>
-          <button onClick ={handleClear}>Poista tiedot</button>
+      <button onClick ={handleClear}>Poista kaikki tiedot</button>
     </div>
-
   )
-
 }
 
-const AllData = ({data}) => {
-  const dates = data.labels
-  const weights = data.datasets[0].data
+const AllData = ({data, setWeightData}) => {
   
+    if(data.length === 0){
+      return;
+    }
+    const dates = data.labels
+    const weights = data.datasets[0].data
+  
+    const handleButtonClick = (date) => {
+      replaceData(date,'',setWeightData)
+    }
+
   return(
     <>
-    
-    
+    {dates.map((date, index) => (
+      <div key={date}>
+        <button  onClick={() => handleButtonClick((date))}>
+          {`Poista tieto päivämäärältä ${(date)}, Paino: ${weights[index]} kg`}
+        </button>
+        <br/>
+    </div>
+      ))}
     </>
   )
 }
 
 const App= () => {
-  const currentDate = new Date();
-  const year = currentDate.getFullYear();
-  const month = (currentDate.getMonth() + 1).toString().padStart(2,'0')
-  const day = currentDate.getDate().toString().padStart(2,'0')
-  const formattedDate = `${day}-${month}-${year}`
+  const dateString = stringifyDate(new Date())
 
   const dataString = localStorage.getItem('weightData')
   const data = dataString ? JSON.parse(dataString) : [];
 
-  
-  const [date, setDate] = useState(formattedDate);
+  const [date, setDate] = useState(dateString);
   const [weight, setWeight] = useState('')
 
   const [weightData, setWeightData] = useState({
-    labels: data.map((data) => (parseDate(data.date))),
+    labels: data.map((data) => ((data.date))),
     datasets: [{
       label:"Paino",
       data: data.map((data)=> data.weight),
@@ -144,26 +140,18 @@ const App= () => {
 
 
   return (
-
     <div className="App">
         <h1>Painonseuranta</h1>
         <h3>Syötä päivämäärä ja paino</h3>
         
-        <DatePicker date={date} setDate={setDate} ></DatePicker>
-        <WeightForm weight={weight} setWeight={setWeight} setWeightData={setWeightData} date={date}></WeightForm>
-        <div className="App">
-          <LineChart chartData={weightData}></LineChart>
-        </div>
-        <ClearButton setWeightData = {setWeightData}></ClearButton>
+        <DatePicker date={date} setDate={setDate}/>
+        <WeightForm weight={weight} setWeight={setWeight} setWeightData = {setWeightData} date={date}/>
+        <LineChart chartData={weightData}/>
         <br/>
-        <AllData data={weightData}></AllData>
-        <div>
-        </div>
-
-
+        <AllData data={weightData} setWeightData={setWeightData}/>
+        <br/>
+        <ClearButton setWeightData = {setWeightData}/>
     </div>
-
-    
   );
 }
